refactor log4j2 sample
