+++ springboot3-samples/logging/log4j2/base/src/main/java/com/alipay/sofa/logging/log4j2/base/facade/SampleService.java
@@ -14,16 +14,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alipay.sofa.web.biz1;
+package com.alipay.sofa.logging.log4j2.base.facade;
 
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
+public interface SampleService {
 
-@SpringBootTest
-class Biz1ApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
+    /**
+     * a simple facade
+     * @return
+     */
+    String service();
 }
