+++ springboot3-samples/logging/log4j2/base/src/main/java/com/alipay/sofa/logging/log4j2/base/impl/SampleServiceImpl.java
@@ -14,13 +14,27 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alipay.sofa.web.base.facade;
+package com.alipay.sofa.logging.log4j2.base.impl;
 
-public interface SampleService {
+import com.alipay.sofa.logging.log4j2.base.facade.SampleService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.ApplicationContext;
+import org.springframework.stereotype.Service;
 
-    /**
-     * a simple facade
-     * @return
-     */
-    String service();
+@Service
+public class SampleServiceImpl implements SampleService {
+    private static Logger      LOGGER = LoggerFactory.getLogger(SampleServiceImpl.class);
+
+    @Autowired
+    private ApplicationContext applicationContext;
+
+    @Override
+    public String service() {
+        String appName = applicationContext.getId();
+
+        LOGGER.info("{} web test: into a service", appName);
+        return "A Sample Service";
+    }
 }
