+++ springboot3-samples/logging/log4j2/base/src/main/java/com/alipay/sofa/logging/log4j2/base/rest/SampleController.java
@@ -14,9 +14,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alipay.sofa.web.base.rest;
+package com.alipay.sofa.logging.log4j2.base.rest;
 
-import com.alipay.sofa.web.base.facade.SampleService;
+import com.alipay.sofa.logging.log4j2.base.facade.SampleService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
