+++ springboot3-samples/logging/log4j2/base/src/main/resources/META-INF/spring/service.xml
@@ -3,6 +3,6 @@
        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
        xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd">
 
-    <bean id="sampleService" class="com.alipay.sofa.web.base.impl.SampleServiceImpl"/>
+    <bean id="sampleService" class="com.alipay.sofa.logging.log4j2.base.impl.SampleServiceImpl"/>
 
 </beans>
