+++ springboot3-samples/logging/log4j2/base/src/test/java/com/alipay/sofa/logging/log4j2/base/BaseApplicationTests.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alipay.sofa.web.base;
+package com.alipay.sofa.logging.log4j2.base;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
