+++ springboot3-samples/logging/log4j2/biz1/src/test/java/com/alipay/sofa/logging/log4j2/biz1/Biz1ApplicationTests.java
@@ -14,27 +14,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alipay.sofa.web.base.impl;
+package com.alipay.sofa.logging.log4j2.biz1;
 
-import com.alipay.sofa.web.base.facade.SampleService;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.ApplicationContext;
-import org.springframework.stereotype.Service;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
 
-@Service
-public class SampleServiceImpl implements SampleService {
-    private static Logger      LOGGER = LoggerFactory.getLogger(SampleServiceImpl.class);
+@SpringBootTest
+class Biz1ApplicationTests {
 
-    @Autowired
-    private ApplicationContext applicationContext;
-
-    @Override
-    public String service() {
-        String appName = applicationContext.getId();
-
-        LOGGER.info("{} web test: into a service", appName);
-        return "A Sample Service";
+    @Test
+    void contextLoads() {
     }
+
 }
