+++ springboot3-samples/logging/log4j2/biz2/pom.xml
@@ -3,10 +3,10 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <groupId>com.alipay.sofa</groupId>
-        <artifactId>springboot3-samples</artifactId>
+        <groupId>com.alipay.sofa.log4j2</groupId>
+        <artifactId>log4j2</artifactId>
         <version>0.0.1-SNAPSHOT</version>
-        <relativePath>../../../pom.xml</relativePath> <!-- lookup parent from repository -->
+        <relativePath>../pom.xml</relativePath> <!-- lookup parent from repository -->
     </parent>
     <groupId>com.alipay.sofa.logging.log4j2</groupId>
     <artifactId>biz2-log4j2</artifactId>
