--- springboot3-samples/logging/log4j2/biz2/src/main/java/com/alipay/sofa/web/biz2/rest/SampleController.java
@@ -1,40 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.alipay.sofa.web.biz2.rest;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.ApplicationContext;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class SampleController {
-    private static Logger      LOGGER = LoggerFactory.getLogger(SampleController.class);
-
-    @Autowired
-    private ApplicationContext applicationContext;
-
-    @RequestMapping(value = "/", method = RequestMethod.GET)
-    public String hello() {
-        String appName = applicationContext.getId();
-        LOGGER.info("{} web test: into sample controller", appName);
-        return String.format("hello to %s deploy", appName);
-    }
-}
